refine the implementation of malachite metric fetcher
