+++ pkg/metaserver/agent/metric/malachite/fetcher_calculate.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package metric
+package malachite
 
 // for those metrics need extra calculation logic,
 // we will put them in a separate file here
